Move additional wizard step into it's own controller
