+++ app/views/works_update_additional/update_additional.html.erb
@@ -1,7 +1,7 @@
 <div class="wizard-area">
   <h1>New Submission</h1>
-  <%= render "wizard_progress", wizard_step: 1 %>
-  <%= render partial: 'form_wizard', locals: { form_submit_url: work_update_additional_path(@work),
+  <%= render "/works_wizard/wizard_progress", wizard_step: 1 %>
+  <%= render partial: '/works_wizard/form_wizard', locals: { form_submit_url: work_update_additional_path(@work),
                                                additional_selected: true, required_selected: false,
                                                required_class: "", additional_class: "show active"} %>
 </div>
